+++ R/data.R
@@ -4,7 +4,7 @@
 #' deflators that can be used to directly compare absolute monetary values across different
 #' LIS and LWS datasets.
 #'
-#' @format A data frame with 7 variables:
+#' @format A data frame with 8 variables:
 #'
 #' @source \url{https://www.lisdatacenter.org/resources/ppp-deflators/}
 "deflators"